backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-core
pinecone
sentence-transformers
pypdf
python-docx
docx2txt
python-dotenv
python-multipart
requests
rank-bm25
numpy